requirements.txt
python-dotenv
PyJWT
djangorestframework
mysqlclient
pymysql
djangorestframework-simplejwt